add test for key generator
