+++ lib/enigma.rb
@@ -7,11 +7,11 @@ class Enigma
   def initialize
   end
 
-  def encrypt(message, key, date)
+  def encrypt(message, key = nil, date = nil)
     Encryptor.new(message, key, date).encrypt
   end
 
-  def decrypt(message, key, date)
+  def decrypt(message, key, date =nil)
     Decryptor.new(message, key, date).decrypt
   end
