+++ spec/enigma_spec.rb
@@ -1,5 +1,6 @@
 require './lib/enigma'
 
+
 RSpec.describe Enigma do
   before :each do
     @enigma = Enigma.new
